Delete list.go and main.go, refactor DockerCompose management.- Refactor DockerCompose management, remove old code
